Fix username differences in watch list. Add python version check. More error handling.
